Move rasterio-specific code to separate module
